Implemented basic display with sample dynamic draw
Initialize method is done, and the engine is drawing a triangle to the
screen, whose top vertex slides to the left and whose colors change.

The triangle is embedded in the thing, so the next thing to do is create a
sprite class and then texture it. The vertex shader should pass the
textures and all thru