Basic sprite implementation - A square!
Sprite can draw a square to the screen, and the vertices of the square
are colored. The colors are a side effect of the current shader, since
there's no texturing yet. The colors stand in for the four corners of
the texutres. I'll rework it to actually show the texture once I get
those loading.

I don't know if this is the best way to work this, but the Sprite class
is implemented such that all the memory for the sprites lives in a
single buffer, since it (at least /seems/ like) the 3DS can only handle
12 buffers at once, according to the citro3d source code. So, the buffer
has to be num_sprites * Sprite::VerticesPerSprite() * sizeof(Vertex)
bytes in linear RAM. Each sprite has 6 vertices, so ptr, ptr + 6, ptr +
12... are the valid vertex buffer sections.