Screen flip and prep shader for textures
The screen's Y axis is upside down, so that the top left corner of the
screen is (0, 0) instead of the bottom left corner. Found the tip in
devkitPro/examples/3ds/gpu/gpusprites

The vertex shader passes the texture coords to the texture unit (even
tho nothing is actually loaded) and forces the vertex color to white.