Basic GameObject and Sprite in place
Sprite loading works, using a compiled-in png and lodepng.

There is now an interface, IGameObject, which will serve as the base for
all things in the game. The Ball class implements that interface, and
displays an image on the screen that bounces around, like the old DVD
player screensavers.

The Engine currently reserves space for 30 sprites, and things which
need sprites can fetch them from the engine. The current Engine instance
is available from the Engine class. There are no thread protections,
since right now, there are no threads.

The Makefile has more flags for lodepng to disable the components I
don't need. Really, that's everything but the allocators and decoder.

I should change the Point class to a Vector class, probably, since I'm
using it more and more as if it were.